+++ protected/humhub/modules/file/resources/js/humhub.file.js
@@ -12,6 +12,7 @@ humhub.module('file', function (module, require, $) {
     var string = util.string;
     var action = require('action');
     var event = require('event');
+
     
     var view = require('ui.view');
 
@@ -51,6 +52,11 @@ humhub.module('file', function (module, require, $) {
             objectModel: this.$.data('upload-model'),
             objectId: this.$.data('upload-model-id')
         };
+
+        if(this.$.data('upload-hide-in-stream')) {
+            data['hideInStream'] = 1;
+        }
+
         return {
             url: this.$.data('url') || this.$.data('upload-url') || module.config.upload.url,
             dropZone: this.getDropZone(),
@@ -227,7 +233,8 @@ humhub.module('file', function (module, require, $) {
             this.callbacks.done(e, response);
         }
 
-        this.fire('humhub:file:uploadEnd', [response]);
+        //deprecated event use uploadEnd
+        this.fire('humhub:file:uploadEnd', [response]);;
         this.fire('uploadEnd', [response]);
     };