+++ protected/humhub/docs/CHANGELOG.md
@@ -5,7 +5,7 @@ HumHub Change Log
 - Enh: Increase profile image size to 800px.
 - Fix #2644 overlapping popup preview image after increasing preview image size (hagalaz)
 - Fix: Button widget child class static instantiation not working 
-- Fix: ModalButton instatiation and added ModalButton::close()
+- Fix: ModalButton instantiation and added ModalButton::close()
 - Fix: Respect `max_file_uploads` setting in UploadInput widget
 - Enh: Include `kartik-v/yii2-widgets`
 - Enh: Added `getAccessRules()` to `humhub/components/Controller`
@@ -19,6 +19,9 @@ HumHub Change Log
 - Fix: Use of JS-Widget internal event object instead of node
 - Enh: Added `Formatter::getDateTimePattern()` and `Formatter::isShowMeridiem()`
 - Fix: Set formatter locale in I18N when changing locale
+- Enh: Added $hideInStream flag for upload component/action for changing `show_in_stream` file flag
+- Enh: Added $showInStream flag for `FilePreview` widget to only include files with certain `show_in_stream` flag
+- Enh: Added `FileManager::findStreamFiles()` for querying files with either given `show_in_stream = 1` or `show_in_stream = 0` flag.
 
 1.2.1 (June 17, 2017)
 - Fix: Invite error in french language